refactor: :recycle: Move the playground UI into separate component
